# swagger.json
{
  "schemes": [],
  "swagger": "2.0",
  "info": {
      "description": "Passontw Auth Service API.",
      "title": "Passontw Auth Service API",
      "termsOfService": "http://swagger.io/terms/",
      "contact": {
          "name": "API Support",
          "url": "http://www.swagger.io/support",
          "email": "support@swagger.io"
      },
      "license": {
          "name": "Apache 2.0",
          "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
      },
      "version": ""
  },
  "host": "",
  "basePath": "/",
  "paths": {
...
      "/api/auth/login": {
          "post": {
              "description": "驗證用戶憑證並返回訪問令牌",
              "consumes": [
                  "application/json"
              ],
              "produces": [
                  "application/json"
              ],
              "tags": [
                  "授權"
              ],
              "summary": "用戶登入",
              "parameters": [
                  {
                      "description": "登入信息",
                      "name": "login",
                      "in": "body",
                      "required": true,
                      "schema": {
                          "$ref": "#/definitions/models.LoginRequest"
                      }
                  }
              ],
              "responses": {
                  "200": {
                      "description": "成功",
                      "schema": {
                          "$ref": "#/definitions/models.LoginResponse"
                      }
                  },
                  "400": {
                      "description": "請求錯誤",
                      "schema": {
                          "$ref": "#/definitions/interfaces.ErrorResponse"
                      }
                  },
                  "401": {
                      "description": "未授權",
                      "schema": {
                          "$ref": "#/definitions/interfaces.ErrorResponse"
                      }
                  },
                  "500": {
                      "description": "服務器錯誤",
                      "schema": {
                          "$ref": "#/definitions/interfaces.ErrorResponse"
                      }
                  }
              }
          }
      },
...
      "/api/v1/users": {
          "post": {
              "description": "創建一個新的用戶帳號",
              "consumes": [
                  "application/json"
              ],
              "produces": [
                  "application/json"
              ],
              "tags": [
                  "users"
              ],
              "summary": "註冊新用戶",
              "parameters": [
                  {
                      "description": "註冊信息",
                      "name": "request",
                      "in": "body",
                      "required": true,
                      "schema": {
                          "$ref": "#/definitions/models.RegisterRequest"
                      }
                  }
              ],
              "responses": {
                  "200": {
                      "description": "OK",
                      "schema": {
                          "$ref": "#/definitions/models.User"
                      }
                  },
                  "400": {
                      "description": "Bad Request",
                      "schema": {
                          "$ref": "#/definitions/interfaces.ErrorResponse"
                      }
                  },
                  "500": {
                      "description": "Internal Server Error",
                      "schema": {
                          "$ref": "#/definitions/interfaces.ErrorResponse"
                      }
                  }
              }
          }
      },
      "/api/v1/auth/login": {
          "post": {
              "description": "使用電子郵件和密碼登入",
              "consumes": [
                  "application/json"
              ],
              "produces": [
                  "application/json"
              ],
              "tags": [
                  "users"
              ],
              "summary": "用戶登入",
              "parameters": [
                  {
                      "description": "登入信息",
                      "name": "request",
                      "in": "body",
                      "required": true,
                      "schema": {
                          "$ref": "#/definitions/models.LoginRequest"
                      }
                  }
              ],
              "responses": {
                  "200": {
                      "description": "OK",
                      "schema": {
                          "$ref": "#/definitions/models.TokenResponse"
                      }
                  },
                  "401": {
                      "description": "Unauthorized",
                      "schema": {
                          "$ref": "#/definitions/interfaces.ErrorResponse"
                      }
                  },
                  "500": {
                      "description": "Internal Server Error",
                      "schema": {
                          "$ref": "#/definitions/interfaces.ErrorResponse"
                      }
                  }
              }
          }
      },
      "/api/v1/users/profile": {
          "get": {
              "security": [
                  {
                      "Bearer": []
                  }
              ],
              "description": "獲取當前登入用戶的個人資料",
              "consumes": [
                  "application/json"
              ],
              "produces": [
                  "application/json"
              ],
              "tags": [
                  "users"
              ],
              "summary": "獲取用戶資料",
              "responses": {
                  "200": {
                      "description": "OK",
                      "schema": {
                          "$ref": "#/definitions/models.UserProfileResponse"
                      }
                  },
                  "401": {
                      "description": "Unauthorized",
                      "schema": {
                          "$ref": "#/definitions/interfaces.ErrorResponse"
                      }
                  },
                  "500": {
                      "description": "Internal Server Error",
                      "schema": {
                          "$ref": "#/definitions/interfaces.ErrorResponse"
                      }
                  }
              }
          },
          "put": {
              "security": [
                  {
                      "Bearer": []
                  }
              ],
              "description": "更新當前登入用戶的個人資料",
              "consumes": [
                  "application/json"
              ],
              "produces": [
                  "application/json"
              ],
              "tags": [
                  "users"
              ],
              "summary": "更新用戶資料",
              "parameters": [
                  {
                      "description": "更新信息",
                      "name": "request",
                      "in": "body",
                      "required": true,
                      "schema": {
                          "$ref": "#/definitions/models.UpdateProfileRequest"
                      }
                  }
              ],
              "responses": {
                  "200": {
                      "description": "OK",
                      "schema": {
                          "$ref": "#/definitions/handler.SuccessResponse"
                      }
                  },
                  "400": {
                      "description": "Bad Request",
                      "schema": {
                          "$ref": "#/definitions/interfaces.ErrorResponse"
                      }
                  },
                  "401": {
                      "description": "Unauthorized",
                      "schema": {
                          "$ref": "#/definitions/interfaces.ErrorResponse"
                      }
                  },
                  "500": {
                      "description": "Internal Server Error",
                      "schema": {
                          "$ref": "#/definitions/interfaces.ErrorResponse"
                      }
                  }
              }
          }
      },
      "/api/v1/users/settings": {
          "put": {
              "security": [
                  {
                      "Bearer": []
                  }
              ],
              "description": "更新當前登入用戶的個人設定",
              "consumes": [
                  "application/json"
              ],
              "produces": [
                  "application/json"
              ],
              "tags": [
                  "users"
              ],
              "summary": "更新用戶設定",
              "parameters": [
                  {
                      "description": "設定信息",
                      "name": "request",
                      "in": "body",
                      "required": true,
                      "schema": {
                          "$ref": "#/definitions/models.AppUpdateSettingsRequest"
                      }
                  }
              ],
              "responses": {
                  "200": {
                      "description": "OK",
                      "schema": {
                          "$ref": "#/definitions/handler.SuccessResponse"
                      }
                  },
                  "400": {
                      "description": "Bad Request",
                      "schema": {
                          "$ref": "#/definitions/interfaces.ErrorResponse"
                      }
                  },
                  "401": {
                      "description": "Unauthorized",
                      "schema": {
                          "$ref": "#/definitions/interfaces.ErrorResponse"
                      }
                  },
                  "500": {
                      "description": "Internal Server Error",
                      "schema": {
                          "$ref": "#/definitions/interfaces.ErrorResponse"
                      }
                  }
              }
          }
      },
...
  },
  "definitions": {
...
      "handler.SuccessResponse": {
          "type": "object",
          "properties": {
              "message": {
                  "type": "string"
              }
          }
      },
      "interfaces.ErrorResponse": {
          "type": "object",
          "properties": {
              "error": {
                  "type": "string"
              }
          }
      },
...
      "models.AppUpdateSettingsRequest": {
          "type": "object",
          "properties": {
              "ai_assistant": {
                  "type": "boolean",
                  "example": true
              },
              "data_collection": {
                  "type": "boolean",
                  "example": true
              },
              "game_recommendation": {
                  "type": "boolean",
                  "example": true
              },
              "high_quality": {
                  "type": "boolean",
                  "example": true
              },
              "music": {
                  "type": "boolean",
                  "example": true
              },
              "sound": {
                  "type": "boolean",
                  "example": true
              },
              "vibration": {
                  "type": "boolean",
                  "example": false
              }
          }
      },
...
      "models.LoginRequest": {
          "type": "object",
          "required": [
              "email",
              "password"
          ],
          "properties": {
              "email": {
                  "type": "string"
              },
              "password": {
                  "type": "string"
              }
          }
      },
      "models.LoginResponse": {
          "type": "object",
          "properties": {
              "expires_in": {
                  "type": "integer",
                  "example": 3600
              },
              "token": {
                  "type": "string",
                  "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
              },
              "token_type": {
                  "type": "string",
                  "example": "Bearer"
              }
          }
      },
...
      "models.RegisterRequest": {
          "type": "object",
          "required": [
              "email",
              "password",
              "username"
          ],
          "properties": {
              "email": {
                  "type": "string"
              },
              "password": {
                  "type": "string",
                  "minLength": 8
              },
              "username": {
                  "type": "string",
                  "maxLength": 50,
                  "minLength": 3
              }
          }
      },
...
      "models.TokenResponse": {
          "type": "object",
          "properties": {
              "token": {
                  "type": "string"
              }
          }
      },
      "models.UpdateProfileRequest": {
          "type": "object",
          "properties": {
              "avatar_url": {
                  "type": "string"
              },
              "username": {
                  "type": "string",
                  "maxLength": 50,
                  "minLength": 3
              }
          }
      },
      "models.User": {
          "type": "object",
          "properties": {
              "auth_provider": {
                  "type": "string"
              },
              "auth_provider_id": {
                  "type": "string"
              },
              "avatar_url": {
                  "type": "string"
              },
              "created_at": {
                  "type": "string"
              },
              "email": {
                  "type": "string"
              },
              "is_active": {
                  "type": "boolean"
              },
              "is_verified": {
                  "type": "boolean"
              },
              "last_login_at": {
                  "type": "string"
              },
              "points": {
                  "type": "integer"
              },
              "role": {
                  "type": "string"
              },
              "updated_at": {
                  "type": "string"
              },
              "user_id": {
                  "type": "string"
              },
              "username": {
                  "type": "string"
              },
              "vip_level": {
                  "type": "integer"
              }
          }
      },
      "models.UserProfileResponse": {
          "type": "object",
          "properties": {
              "settings": {
                  "$ref": "#/definitions/models.UserSettings"
              },
              "user": {
                  "$ref": "#/definitions/models.User"
              },
              "wallet": {
                  "$ref": "#/definitions/models.UserWallet"
              }
          }
      },
      "models.UserSettings": {
          "type": "object",
          "properties": {
              "ai_assistant": {
                  "type": "boolean"
              },
              "data_collection": {
                  "type": "boolean"
              },
              "game_recommendation": {
                  "type": "boolean"
              },
              "high_quality": {
                  "type": "boolean"
              },
              "music": {
                  "type": "boolean"
              },
              "sound": {
                  "type": "boolean"
              },
              "updated_at": {
                  "type": "string"
              },
              "user_id": {
                  "type": "string"
              },
              "vibration": {
                  "type": "boolean"
              }
          }
      },
      "models.UserWallet": {
          "type": "object",
          "properties": {
              "balance": {
                  "type": "number"
              },
              "created_at": {
                  "type": "string"
              },
              "total_bet": {
                  "type": "number"
              },
              "total_deposit": {
                  "type": "number"
              },
              "total_win": {
                  "type": "number"
              },
              "total_withdraw": {
                  "type": "number"
              },
              "updated_at": {
                  "type": "string"
              },
              "user_id": {
                  "type": "string"
              },
              "wallet_id": {
                  "type": "string"
              }
          }
      },
...
  },
  "securityDefinitions": {
      "Bearer": {
          "description": "Type \"Bearer\" followed by a space and JWT token.",
          "type": "apiKey",
          "name": "Authorization",
          "in": "header"
      }
  }
}